Refactor JNI for VAD + ASR
