+++ android/SherpaOnnxVadAsr/app/src/main/java/com/k2fsa/sherpa/onnx/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.k2fsa.sherpa.onnx
+package com.k2fsa.sherpa.onnx.vad.asr
 
 import android.Manifest
 import android.content.pm.PackageManager
@@ -13,6 +13,13 @@ import android.widget.Button
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
+import com.k2fsa.sherpa.onnx.OfflineRecognizer
+import com.k2fsa.sherpa.onnx.OfflineRecognizerConfig
+import com.k2fsa.sherpa.onnx.R
+import com.k2fsa.sherpa.onnx.Vad
+import com.k2fsa.sherpa.onnx.getFeatureConfig
+import com.k2fsa.sherpa.onnx.getOfflineModelConfig
+import com.k2fsa.sherpa.onnx.getVadModelConfig
 import kotlin.concurrent.thread
 
 
@@ -40,7 +47,7 @@ class MainActivity : AppCompatActivity() {
     private val permissions: Array<String> = arrayOf(Manifest.permission.RECORD_AUDIO)
 
     // Non-streaming ASR
-    private lateinit var offlineRecognizer: SherpaOnnxOffline
+    private lateinit var offlineRecognizer: OfflineRecognizer
 
     private var idx: Int = 0
     private var lastText: String = ""
@@ -201,13 +208,17 @@ class MainActivity : AppCompatActivity() {
             modelConfig = getOfflineModelConfig(type = secondType)!!,
         )
 
-        offlineRecognizer = SherpaOnnxOffline(
+        offlineRecognizer = OfflineRecognizer(
             assetManager = application.assets,
             config = config,
         )
     }
 
     private fun runSecondPass(samples: FloatArray): String {
-        return offlineRecognizer.decode(samples, sampleRateInHz)
+        val stream = offlineRecognizer.createStream()
+        stream.acceptWaveform(samples, sampleRateInHz)
+        offlineRecognizer.decode(stream)
+        val result = offlineRecognizer.getResult(stream)
+        return result.text
     }
 }
